% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Start a Shiny-App as GUI for all workflows}
\usage{
app()
}
\value{
A Shiny-App object. This will effectively launch a browser and start
  the app on local port 7462.
}
\description{
Start a Shiny-App that allows to upload raw data, process selectively and
analyze different methods of ratio calculation between two intensity traces.
}
\details{
The app contains context specific help accessible during run-time
    via links.
}
