% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_analyte_mass_as_element}
\alias{calc_analyte_mass_as_element}
\title{calc_analyte_mass_as_element.}
\usage{
calc_analyte_mass_as_element(
  R_m,
  K,
  Asp_iso1,
  Asp_iso2,
  As_iso1,
  As_iso2,
  N_sp
)
}
\arguments{
\item{R_m}{A vector of isotope ratios.}

\item{K}{Massbias K.}

\item{Asp_iso1}{Abundance of the spike isotope in the spike.}

\item{Asp_iso2}{Abundance of the sample isotope in the spike.}

\item{As_iso1}{Natural abundance of the spike isotope.}

\item{As_iso2}{Natural abundance of the sample isotope.}

\item{N_sp}{N_sp * Molar mass of sample.}
}
\description{
calc_analyte_mass_as_element.
}
