% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_cali_mod}
\alias{calc_cali_mod}
\title{Calculates the calibration model}
\usage{
calc_cali_mod(
  df,
  wf = c("ExtCal", "ExtGasCal", "oIDMS"),
  ExtCal_unit = c("pg", "ng", "µg"),
  ExtGasCal_unit = c("nL/min", "µL/min", "mL/min")
)
}
\arguments{
\item{df}{Data.frame with two columns.}

\item{wf}{Calibration method/Workflow (used to make assumptions regarding units in df).}

\item{ExtCal_unit}{The measurement unit of the ExtCal workflow to correctly format the output.}

\item{ExtGasCal_unit}{The measurement unit of the ExtGasCal workflow to correctly format the output.}
}
\value{
A data.frame containing slope and intercept with errors and R square.
}
\description{
\code{calc_cali_mod} will provide calibration results based on a linear
    regression model.
}
\details{
Calculates the calibration model.

A calibration curve is provided by a linear fit of peak areas or mean
    signal intensities against analyte masses or gas flows from a data.frame
    containing at least two entries in rows.
}
\examples{
df <- data.frame(x=1:5, y=sort(runif(5)))
calc_cali_mod(df = df, wf = "ExtCal")
calc_cali_mod(df = df, wf = "ExtGasCal")
calc_cali_mod(df = df, wf = "oIDMS")
}
