% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_massflow}
\alias{calc_massflow}
\title{calc_massflow.}
\usage{
calc_massflow(
  x,
  n_trans = 16.64236,
  As_iso1 = 7.68,
  As_iso2 = 4.36,
  Asp_iso1 = 91.06,
  Asp_iso2 = 0.08,
  V_fl = 0.0075,
  c_sp = 19581.71,
  DF = 20
)
}
\arguments{
\item{x}{Input data.}

\item{n_trans}{n_trans.}

\item{As_iso1}{Natural abundance of the spike isotope.}

\item{As_iso2}{Natural abundance of the sample isotope.}

\item{Asp_iso1}{Abundance of the spike isotope in the spike.}

\item{Asp_iso2}{Abundance of the sample isotope in the spike.}

\item{V_fl}{Sample inlet flow in mL/min.}

\item{c_sp}{Concentration of spike isotope in spike stock solution.}

\item{DF}{Dilution factor of spike solution.}
}
\value{
A numeric vector.
}
\description{
Transforms one column in a data.frame with respect to parameters.
}
\details{
xx
}
\examples{
calc_massflow(1:10)

}
