% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_N_sp}
\alias{calc_N_sp}
\title{calc_N_sp.}
\usage{
calc_N_sp(c_sp, V_sp, VF1, M_sp, M_sa)
}
\arguments{
\item{c_sp}{Concentration of spike isotope in spike stock solution [mg/L].}

\item{V_sp}{Volume of spike solution [µL].}

\item{VF1}{Dilution factor of spike solution.}

\item{M_sp}{Molar mass of spike.}

\item{M_sa}{Molar mass of sample.}
}
\description{
calc_N_sp.
}
