% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_transeff}
\alias{calc_transeff}
\title{calc_transeff}
\usage{
calc_transeff(
  data,
  int_col,
  LFD = 100,
  cali_slope = 1,
  V_fl,
  part_mat = c("Au", "Ag"),
  dia_part
)
}
\arguments{
\item{data}{A data.frame containing at least two columns.}

\item{int_col}{Intensity column.}

\item{LFD}{Intensity limit for the detection of particle signals.}

\item{cali_slope}{Calibration lm slope result for ionic standards.}

\item{V_fl}{Sample inlet flow in mL/min.}

\item{part_mat}{Particle material.}

\item{dia_part}{Particle diameter in nm.}
}
\value{
A data.frame containing single particle data.
}
\description{
The transport efficiency will be calculated based on the particle
    size approach from single particle-ICP-MS data.
}
\details{
Determination of the transport efficiency from measurements with
    gold or silver nano particle standards of known size.
}
\examples{
spion_imp <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_ionic"]]

head(spion_imp[[1]])
names(spion_imp)
sp_cali <- ldply_base(1:length(spion_imp), function(i) {
  get_peakdata(spion_imp[[i]], "197Au", PPmethod = "mean_signal", peak_start = 0.003, peak_end = 60)
})
conc_ion <- c(20, 50, 100, 200, 500)
fac <- 1.661 * 0.01104347 * 12/44
sp_cali <- tab_cali(sp_cali, wf = "oIDMS", std_info = conc_ion, gas_flow = 49.6512, fac = fac)

(cali_slope <- calc_cali_mod(df = sp_cali[,c(5,4)], wf = "oIDMS")[,1])

sp_data <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_particle"]][[1]]
# filter time if required
sp_data_flt <- sp_data[sp_data[,"Time"]<60,]

(n_trans <- calc_transeff(sp_data, int_col = "197Au", LFD = 20000, cali_slope,
  V_fl = 0.0075, part_mat = c("Au", "Ag"), dia_part = 60))
}
