% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{ETVapp_testdata}
\alias{ETVapp_testdata}
\title{ETVapp_testdata.}
\format{
A named list of length 4 (respective of the 4 implemented workflows),
    each containing sublists of blanks, calibration data and measured samples
    imported by `import_data()`.
}
\source{
vera.scharek@bam.de
}
\usage{
data(ETVapp_testdata)
}
\description{
ETVapp_testdata.
}
\keyword{datasets}
