% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{extract_unique_number}
\alias{extract_unique_number}
\title{extract_unique_number.}
\usage{
extract_unique_number(x)
}
\arguments{
\item{x}{Charcter vector, i.e. of file names.}
}
\description{
Often numeric information is stored in character vectors like
    file names. This function will extract such information, focusing on
    numeric values which are different between files.
}
\examples{
x <- paste("File123_conc", c(3, 21, 101), "endtext", sep="_")
extract_unique_number(x=x)
}
