% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isoratio.R
\name{get_isoratio}
\alias{get_isoratio}
\title{Calculate an isotope ratio based on peak areas.}
\usage{
get_isoratio(
  data,
  iso1_col,
  iso2_col,
  PPmethod = c("Peak (height)", "Peak (manual)"),
  peak_start,
  peak_end,
  minpeakheight = 1000,
  BLmethod = c("modpolyfit", "none"),
  deg = 1,
  cf = 50,
  fl = 5,
  simplify = TRUE
)
}
\arguments{
\item{data}{A data.frame containing at least two columns or a list of such
data.frames in which case the result will be a list of data frames as well.}

\item{iso1_col}{Spike isotope.}

\item{iso2_col}{Sample isotope.}

\item{PPmethod}{Peak picking method.}

\item{peak_start}{Value which is taken as peak start point, when manual peak picking is chosen.}

\item{peak_end}{Value which is taken as peak end point, when manual peak picking is chosen.}

\item{minpeakheight}{A threshold value for peak picking when "Peak (height)" is choosen as an option.}

\item{BLmethod}{Method for baseline correction.}

\item{deg}{Degree of polynomial for baseline correction.}

\item{cf}{A correction value for cutting the area around the detected peak.}

\item{fl}{Filter length of smoothing function, has to be odd integer >=3.}

\item{simplify}{In case that data is a list: shall result table be combined to a data.frame?}
}
\value{
A data.frame or a list of data.frames when data is a list itself.
}
\description{
Calculate an isotope ratio based on peak areas.
}
\details{
Time-resolved ICP-MS data will be integrated and optionally baseline
    corrected. Peak boundaries will be checked regarding the similarity
    between the isotopes.
}
\examples{
# for sample measurements
td <- ETVapp::ETVapp_testdata[["IDMS"]][["Samples"]]
get_isoratio(
  td, iso1_col = "113Cd", iso2_col = "111Cd", PPmethod = "Peak (manual)",
  peak_start = 72, peak_end = 132
)

}
