% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(file_path, simplify = TRUE)
}
\arguments{
\item{file_path}{Valid file path.}

\item{simplify}{Use FALSE to ensure that names list is returned also for single file as input.}
}
\value{
A data.frame containing numeric columns or a list of such data.frames
    in case that file_path is a vector.
}
\description{
\code{import_data} will import time resolved ICP-MS data in
    various file formats from different instrument types.
}
\details{
The function works for .csv, .TXT, and .exp files. Codes are
    provided to transform raw data from common ICP-MS instruments to a
    data.frame with a time column and additional columns for the acquired
    intensities. You may check why data import of your files fails in the app
    on this function and potentially extend it to handle your files.
}
\examples{
pth <- system.file(package = "ETVapp", "extdata", 'fmt')
imp <- import_data(dir(pth, full.names = TRUE))
lapply(imp, head)

import_data("not_existent.file")

\dontrun{
wfs <- c("ExtCal", "ExtGasCal", "IDMS", "oIDMS")
ETVapp_testdata <- setNames(lapply(wfs, function(wf) {
  subflds <- dir(system.file(package = "ETVapp", "extdata", wf))
  setNames(lapply(subflds, function(sf) {
  fls <- dir(system.file(package = "ETVapp", "extdata", wf, sf), full.names = TRUE)
    import_data(fls, simplify = FALSE)
  }), subfld)
}), wfs)
usethis::use_data(ETVapp_testdata)
}

}
