% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{ldply_base}
\alias{ldply_base}
\title{Base-R ldply implementation}
\usage{
ldply_base(.data, .fun = identity)
}
\arguments{
\item{.data}{list.}

\item{.fun}{fun.}
}
\description{
A base R only implementation of plyr::ldply.
}
