% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plots.R
\name{plot_particle_size_distribution}
\alias{plot_particle_size_distribution}
\title{plot_particle_size_distribution.}
\usage{
plot_particle_size_distribution(
  x,
  cali_slope = 1488,
  V_fl = 0.0075,
  part_mat = c("Au", "Ag"),
  dia_part = 60,
  LFD = 20000
)
}
\arguments{
\item{x}{A data.frame containing at least two columns.}

\item{cali_slope}{Calibration lm slope result for ionic standards.}

\item{V_fl}{Sample inlet flow in mL/min.}

\item{part_mat}{Particle material.}

\item{dia_part}{Particle diameter in nm.}

\item{LFD}{Intensity limit for the detection of particle signals.}
}
\value{
A data.frame containing single particle data (invisible) and a plot.
}
\description{
plot_particle_size_distribution.
}
\details{
Determination of the transport efficiency from measurements with
    gold or silver nano particle standards.
}
\examples{
sp_data <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_particle"]][[1]]
head(plot_particle_size_distribution(x = sp_data))

}
