% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plots.R
\name{plot_spec}
\alias{plot_spec}
\title{Spectra and peaks plot.}
\usage{
plot_spec(
  opt = "",
  xrng = NULL,
  yrng = NULL,
  mi_spec = NULL,
  c1 = "13C",
  xlab = paste0("Time [", "s", "]"),
  ylab = "Intensity [cps]",
  T_prog = NULL,
  s_focus = NULL,
  pks = NULL,
  BLmethod = "none",
  sel_pk = NULL
)
}
\arguments{
\item{opt}{Vector of keywords to hide/show specific elements of the plot.}

\item{xrng}{Numeric vector of length 2 specifying plotting range for x.}

\item{yrng}{Numeric vector of length 2 specifying plotting range for y.}

\item{mi_spec}{Spectra (list of data.frame objects).}

\item{c1}{Main isotope.}

\item{xlab}{Character to be used for bottom axis labeling.}

\item{ylab}{Character to be used for left axis labeling.}

\item{T_prog}{Temperature program (a data frame with columns Time and Temp).}

\item{s_focus}{Index of sample within focus.}

\item{pks}{Data frame of peaks with columns 'Sample', 'Peak ID', 'Scan start' and 'Scan end'.}

\item{BLmethod}{Hide or show baseline.}

\item{sel_pk}{Selected peak as numeric index of length one.}
}
\value{
An annotated plot of one to several spectra.
}
\description{
Generate an annotated plot of one to several spectra.
}
\examples{
if (interactive()) {
  wf <- "ExtGasCal"
  c1 <- "13C"
  td <- ETVapp::ETVapp_testdata[[wf]][["Samples"]]
  pro_data <- process_data(td, wf=wf, c1=c1, fl=9)
  plot_spec(mi_spec = pro_data, opt = "overlay_legend")
  peak_data <- cbind(
    "Sample"=1:length(pro_data),
    get_peakdata(pro_data, int_col = c1, minpeakheight = 10^6)
  )
  T_prog <- data.frame(
    "Time"=seq(min(pro_data[[1]][,"Time"]), max(pro_data[[1]][,"Time"]), length.out=7),
    "Temp"=c(20,20,50,50,150,170,170)
  )
  plot_spec(
    mi_spec = pro_data, opt=c("overlay_pb", "overlay_Temp"), BLmethod = "modpolyfit",
    T_prog = T_prog, pks = peak_data, s_focus = 2
  )
}
}
