% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{process_data}
\usage{
process_data(
  data,
  wf = c("ExtCal", "ExtGasCal", "IDMS", "oIDMS"),
  c1,
  c2 = NULL,
  fl = NULL,
  amend = FALSE
)
}
\arguments{
\item{data}{Data.frame with at least two numeric columns.}

\item{wf}{Calibration method/Workflow.}

\item{c1}{Column name of the intensity column to be used (ExtCal) or of the spike isotope (IDMS).}

\item{c2}{Column name of the intensity column to be used as internal standard for Argon correction (ExtCal) or sample isotope (IDMS).}

\item{fl}{Filter length, has to be odd.}

\item{amend}{Set TRUE to amend transformed columns instead of replacing them.}
}
\value{
A data.frame with two columns.
}
\description{
Correct a numeric values in a data.frame column with respect to
    a column containing standard values and applies Savitzky-Golay smoothing.
}
\details{
Additional details...
}
\examples{
imp <- ETVapp::ETVapp_testdata[['ExtGasCal']][['Samples']][[1]]
plot(imp[,c("Time","13C")], type="l")
pro_data <- process_data(imp, c1 = "13C", c2 = "80Se", fl = 5)
head(pro_data)
lines(pro_data, col=3)

head(process_data(imp, c1 = "13C", c2 = "80Se", fl = 5, amend = TRUE))

# test all error messages
\dontrun{
  colnames(imp)[1] <- "TIME"
  head(process_data(imp, c1 = "122Sn", c2 = "test", fl = 2))
}

}
