% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{str_sort_num}
\alias{str_sort_num}
\title{str_sort_num}
\usage{
str_sort_num(x)
}
\arguments{
\item{x}{character vector.}
}
\description{
str_sort_num
}
\examples{
x <- paste0("File", c(20, 1, 9, 11, 32, 100))
str_sort_num(x)
x <- c("Z1", x, "a", "Z10")
str_sort_num(x)
x <- c("b1_3", "b2_1_3", x)
str_sort_num(x)
}
