% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cali.R
\name{tab_cali}
\alias{tab_cali}
\title{tab_cali.}
\usage{
tab_cali(
  peak_data,
  wf = c("ExtCal", "ExtGasCal", "oIDMS"),
  ExtCal_unit = c("pg", "ng", "µg"),
  ExtGasCal_unit = c("nL/min", "µL/min", "mL/min"),
  std_info = 0,
  fac = 0
)
}
\arguments{
\item{peak_data}{Data.frame containing peak information.}

\item{wf}{Calibration method/Workflow.}

\item{ExtCal_unit}{The measurement unit of the ExtCal workflow to correctly format the output.}

\item{ExtGasCal_unit}{The measurement unit of the ExtGasCal workflow to correctly format the output.}

\item{std_info}{A numeric value giving the analyte mass (ExtCal), the concentration of an ionic standard solution in µg/L (oIDMS) or the gas flow of calibration gas (ExtGasCal).}

\item{fac}{A factor to convert the gas flow (gas_density x mass_fraction x mass_percentage).}
}
\value{
A data.frame with the standard no, the isotope, the peak/signal boundaries in seconds, the peak area in counts/mean signal in cps and information for the
analyte content in the calibration standard.
}
\description{
\code{tab_cali} will transfer standard information to a data.frame containing peak/signal data.
}
\details{
Selecting "ExtGasCal" enables the input of a conversion factor to calculate the gas flows. A conversion factor (mL/min to µL/s) is implemented in the function.
}
\examples{
# import all cali files into a list from example data folder
cali_imp <- ETVapp::ETVapp_testdata[["ExtCal"]][["Cali"]]

# process cali files (and generate plots)
cali_pro <- lapply(cali_imp, function(x) {
  process_data(x, c1 = "157", fl = 7)
})
par(mfrow=grDevices::n2mfrow(nr.plots = length(cali_pro)))
for (i in 1:length(cali_pro)) plot(cali_pro[[i]][,1:2], type="l")

# get cali peaks and combine
peak_start <- rep(145, length(cali_pro))
peak_end <- seq(180, 230, length.out=length(cali_pro))
cali_pks <- ldply_base(1:length(cali_pro), function(i) {
  pk <- get_peakdata(cali_pro[[i]], PPmethod = "Peak (manual)", int_col = "157",
    peak_start = peak_start[i], peak_end = peak_end[i])
})
tab_cali(peak_data = cali_pks, wf = "ExtCal", std_info = seq(0,50,10))

# check if unit specification works
si <- seq(0,50,10)
tab_cali(peak_data = cali_pks, wf = "ExtCal", ExtCal_unit = "ng", std_info = si)
tab_cali(peak_data = cali_pks, wf = "ExtGasCal", ExtGasCal_unit = "\u00b5L/min", std_info = si)

# import and process cali data of `oIDMS` workflow
spion_imp <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_ionic"]]
cali_pks <- ldply_base(1:length(spion_imp), function(i) {
  get_peakdata(spion_imp[[i]][,c("Time", "197Au")], int_col = "197Au",
    PPmethod = "mean signal", peak_start = 0.003, peak_end = 60)
})
tab_cali(peak_data = cali_pks, wf = "oIDMS", std_info = c(20, 50, 100, 200, 500))

}
