% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_LOX.R
\name{tab_LOX}
\alias{tab_LOX}
\title{Tabulate LOD/LOQ data}
\usage{
tab_LOX(
  x,
  cali_slope = 1,
  wf = c("ExtCal", "ExtGasCal", "IDMS", "oIDMS"),
  ExtCal_unit = c("pg", "ng", "µg"),
  ExtGasCal_unit = c("nL/min", "µL/min", "mL/min"),
  c_sp_unit = c("µg/L", "mg/L", "g/L"),
  mass_fraction2 = 1,
  sample_mass = 1
)
}
\arguments{
\item{x}{Variable containing the numeric data to infer SD (either peak area (for ExtCal/ExtGasCal) or analyte mass (for IDMS/oIDMS)).}

\item{cali_slope}{A slope from the calibration data.}

\item{wf}{Calibration method/Workflow.}

\item{ExtCal_unit}{The measurement unit of the ExtCal workflow to correctly format the output.}

\item{ExtGasCal_unit}{The measurement unit of the ExtGasCal_unit workflow to correctly format the output.}

\item{c_sp_unit}{The measurement unit of the "IDMS", "oIDMS" workflow to correctly format the output.}

\item{mass_fraction2}{Mass fraction of the analyte element in the analyte component.}

\item{sample_mass}{Sample mass.}
}
\value{
A data.frame.
}
\description{
Estimates the limit of detection (LOD) and quantification (LOQ) and returns
a data.frame.
}
\details{
The LOD and LOQ will be calculated as three and ten times the
    standard deviation of at least three peak areas divided by the slope
    of a linear calibration curve (for workflow ExtCal/ExtGasCal). Results are provided
    as element, analyte and relative to a sample mass.
}
\examples{
tab_LOX(x = runif(2))
tab_LOX(x = runif(5), wf = "IDMS")
tab_LOX(x = runif(10), mass_fraction2 = 0.5, sample_mass = 3, ExtCal_unit = "ng")

}
