% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{calc_cali_mod}
\alias{calc_cali_mod}
\title{calc_cali_mod.}
\usage{
calc_cali_mod(df, wf = c("ExtCal", "ExtGasCal", "oIDMS"))
}
\arguments{
\item{df}{Data.frame with two columns.}

\item{wf}{Calibration method/Workflow (used to make assumptions regarding units in df).}
}
\value{
A data.frame containing slope and intercept with errors and R square.
}
\description{
\code{calc_cali_mod} will provide calibration results based on linear
    regression model.
}
\details{
A calibration curve is provided by a linear fit of peak areas or mean
    signal intensities against analyte masses or gas flows from a data.frame
    containing at least two entries in rows.
}
\examples{
df <- data.frame(x=1:5, y=sort(runif(5)))
calc_cali_mod(df = df, wf = "ExtCal")
calc_cali_mod(df = df, wf = "ExtGasCal")
calc_cali_mod(df = df, wf = "oIDMS")
}
