% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{correct_ratio}
\alias{correct_ratio}
\title{correct_ratio.}
\usage{
correct_ratio(x, K = 1, As_iso1 = 7.68, As_iso2 = 4.63)
}
\arguments{
\item{x}{Numeric vector of mass ratios (R_m).}

\item{K}{Massbias K.}

\item{As_iso1}{Natural abundance of the spike isotope.}

\item{As_iso2}{Natural abundance of the sample isotope.}
}
\description{
correct_ratio.
}
