% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isoratio.R
\name{get_isoratio}
\alias{get_isoratio}
\title{get_isoratio}
\usage{
get_isoratio(
  data,
  iso1_col,
  iso2_col,
  PPmethod = c("Peak (height)", "Peak (manual)"),
  peak_start,
  peak_end,
  minpeakheight = 1000,
  BLmethod = c("modpolyfit", "none"),
  deg = 1,
  cf = 50,
  fl = 5
)
}
\arguments{
\item{data}{A data.frame containing at least two columns.}

\item{iso1_col}{Spike isotope.}

\item{iso2_col}{Sample isotope.}

\item{PPmethod}{Peak picking method.}

\item{peak_start}{Value which is taken as peak start point, when manual peakpicking is choosen.}

\item{peak_end}{Value which is taken as peak end point, when manual peakpicking is choosen.}

\item{minpeakheight}{A threshold value for peakpicking via peakheight.}

\item{BLmethod}{Method for baseline correction.}

\item{deg}{Degree of polynomial for baseline correction.}

\item{cf}{A correction value for cutting the area around the detected peak.}

\item{fl}{fl.}
}
\value{
A data.frame.
}
\description{
\code{get_isoratio} will be calculate an isotope ratio based on peak areas.
}
\details{
Time-resolved ICP-MS data will be integration and optionally baseline corrected.
}
\examples{
# for sample measurements
lapply(ETVapp::ETVapp_testdata[["IDMS"]], function(x) {
  print(ldply_base(1:length(x), function(i) {
    get_isoratio(
     x[[i]], iso1_col = "113Cd", iso2_col = "111Cd", PPmethod = "Peak (manual)",
     peak_start = 72, peak_end = 132
    )
  }))
})
}
