% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{laply_base}
\alias{laply_base}
\title{Base-R laply implementation}
\usage{
laply_base(.data, .fun, ..., .drop = TRUE)
}
\arguments{
\item{.data}{list.}

\item{.fun}{fun.}

\item{...}{additional parameters to lapply.}

\item{.drop}{Avoid dropping dimensionality.}
}
\description{
A base R only implementation of plyr::laply.
}
\examples{
lst <- list(a = 1:3, b = 4:6, c = 7:9)
laply_base(lst, function(x) x[1:2], .drop = FALSE)
#plyr::laply(lst, function(x) x[1:2], .drop = FALSE)

}
