% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{ldply_base}
\alias{ldply_base}
\title{ldply_base}
\usage{
ldply_base(.data, .fun = identity)
}
\arguments{
\item{.data}{list.}

\item{.fun}{fun.}
}
\description{
exported functions that solve minor problems
}
\details{
# function reimplementations (to avoid dependencies)
- ldply_base
- str_sort_num

perform simple calculations
- calc_N_sp
- calc_analyte_mass_as_element
- correct_ratio
- calc_massbias
- calc_massflow
- calc_cali_mod
- calc_transeff

imported from IsoCor
- get_iso_amu
- help_the_user
- get_spectrum
- spec_pre_process
}
