% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plots.R
\name{plot_particle_diameter}
\alias{plot_particle_diameter}
\title{plot_particle_diameter}
\usage{
plot_particle_diameter(x, LFD = 20000)
}
\arguments{
\item{x}{A data.frame containing at least two columns.}

\item{LFD}{Intensity limit for the detection of particle signals.}
}
\value{
A data.frame containing single particle data (invisible) and a plot.
}
\description{
Plots the signal distribution of a single particle-ICP-MS measurement.
}
\details{
Determination of the intensity limit for the differentiation between particle and background signals.
}
\examples{
sp_data <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_particle"]][[1]]
plot_particle_diameter(x = sp_data[,2])

}
