% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plots.R
\name{plot_signal_distribution}
\alias{plot_signal_distribution}
\title{plot_signal_distribution}
\usage{
plot_signal_distribution(x, LFD = 20000, style = c("hist", "counts"))
}
\arguments{
\item{x}{A data.frame containing at least two columns.}

\item{LFD}{Intensity limit for the detection of particle signals.}

\item{style}{Choose plotting style}
}
\value{
A data.frame containing single particle data (invisible) and a plot.
}
\description{
Plots the signal distribution of a single particle-ICP-MS measurement.
}
\details{
Determination of the intensity limit for the differentiation between particle and background signals.
}
\examples{
sp_data <- ETVapp::ETVapp_testdata[["oIDMS"]][["sp_particle"]][[1]]
plot_signal_distribution(x = sp_data[,2])
plot_signal_distribution(x = sp_data[,2], style="counts")
}
