% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_result.R
\name{tab_result}
\alias{tab_result}
\title{tab_result}
\usage{
tab_result(
  peak_data,
  wf = c("ExtCal", "ExtGasCal", "IDMS", "oIDMS"),
  a = 0,
  b = 1,
  K = 1,
  amae = 1,
  mass_fraction2 = 1,
  sample_mass = 1
)
}
\arguments{
\item{peak_data}{Data.frame containing peak information.}

\item{wf}{Calibration method/Workflow.}

\item{a}{Intercept from a calibration model (for ExtCal).}

\item{b}{Slope from a calibration model (for ExtCal).}

\item{K}{K (for IMDS).}

\item{amae}{Analyte mass as element  (for IMDS).}

\item{mass_fraction2}{Mass fraction of the analyte element in the analyte component.}

\item{sample_mass}{Sample mass in [mg].}
}
\value{
A data.frame.
}
\description{
\code{tab_result} computes result data from peak data or a data.frame containing at least two numeric columns.
}
\details{
Result data table for data of the type "sample". Depending on the selected method, the analyte mass will be
calculated from the peak area and calibration data which are provided in data.frames. For an external gas calibration, values of the
intensity column from a data.frame containing at least two columns are transferred into mass flow data using calibration data from a
data.frame. The analyte mass will be determined via peak integration using the function \code{get_peakdata}. Result data will be provided
for the analyte component itself and as element, as well as content value related to the sample mass.
}
